+++ espflash/src/flasher/mod.rs
@@ -15,8 +15,6 @@ use md5::{Digest, Md5};
 #[cfg(feature = "serialport")]
 use object::{Endianness, read::elf::ElfFile32 as ElfFile};
 use serde::{Deserialize, Serialize};
-#[cfg(feature = "serialport")]
-use serialport::UsbPortInfo;
 use strum::{Display, EnumIter, IntoEnumIterator, VariantNames};
 
 #[cfg(feature = "serialport")]
@@ -31,9 +29,8 @@ use crate::{
 use crate::{
     connection::{
         Connection,
-        Port,
         command::{Command, CommandType},
-        reset::{ResetAfterOperation, ResetBeforeOperation},
+        reset::ResetBeforeOperation,
     },
     error::{ConnectionError, ResultExt as _},
     flasher::stubs::{
@@ -643,31 +640,26 @@ pub struct Flasher {
 #[cfg(feature = "serialport")]
 impl Flasher {
     /// The serial port's baud rate should be 115_200 to connect. After
-    /// connecting, Flasher will change the baud rate to the `speed`
+    /// connecting, Flasher will change the baud rate to the `baud`
     /// parameter.
-    #[allow(clippy::too_many_arguments)]
     pub fn connect(
-        serial: Port,
-        port_info: UsbPortInfo,
-        speed: Option<u32>,
+        mut connection: Connection,
         use_stub: bool,
         verify: bool,
         skip: bool,
         chip: Option<Chip>,
-        after_operation: ResetAfterOperation,
-        before_operation: ResetBeforeOperation,
     ) -> Result<Self, Error> {
-        // Establish a connection to the device using the default baud rate of 115,200
-        // and timeout of 3 seconds.
-        let mut connection = Connection::new(serial, port_info, after_operation, before_operation);
+        // The connection should already be established with the device using the
+        // default baud rate of 115,200 and timeout of 3 seconds.
         connection.begin()?;
         connection.set_timeout(DEFAULT_TIMEOUT)?;
 
         detect_sdm(&mut connection);
 
-        let detected_chip = if before_operation != ResetBeforeOperation::NoResetNoSync {
+        let detected_chip = if connection.before_operation() != ResetBeforeOperation::NoResetNoSync
+        {
             // Detect which chip we are connected to.
-            let detected_chip = detect_chip(&mut connection, use_stub)?;
+            let detected_chip = connection.detect_chip(use_stub)?;
             if let Some(chip) = chip {
                 if chip != detected_chip {
                     return Err(Error::ChipMismatch(
@@ -676,9 +668,10 @@ impl Flasher {
                     ));
                 }
             }
-
             detected_chip
-        } else if before_operation == ResetBeforeOperation::NoResetNoSync && chip.is_some() {
+        } else if connection.before_operation() == ResetBeforeOperation::NoResetNoSync
+            && chip.is_some()
+        {
             chip.unwrap()
         } else {
             return Err(Error::ChipNotProvided);
@@ -694,7 +687,7 @@ impl Flasher {
             skip,
         };
 
-        if before_operation == ResetBeforeOperation::NoResetNoSync {
+        if flasher.connection.before_operation() == ResetBeforeOperation::NoResetNoSync {
             return Ok(flasher);
         }
 
@@ -713,11 +706,9 @@ impl Flasher {
 
         // Now that we have established a connection and detected the chip and flash
         // size, we can set the baud rate of the connection to the configured value.
-        if let Some(baud) = speed {
-            if baud > 115_200 {
-                warn!("Setting baud rate higher than 115,200 can cause issues");
-                flasher.change_baud(baud)?;
-            }
+        if flasher.connection.baud() > 115_200 {
+            warn!("Setting baud rate higher than 115,200 can cause issues");
+            flasher.change_baud(flasher.connection.baud())?;
         }
 
         Ok(flasher)
@@ -790,7 +781,7 @@ impl Flasher {
         }?;
 
         // Re-detect chip to check stub is up
-        let chip = detect_chip(&mut self.connection, self.use_stub)?;
+        let chip = self.connection.detect_chip(self.use_stub)?;
         debug!("Re-detected chip: {:?}", chip);
 
         Ok(())
@@ -1137,11 +1128,11 @@ impl Flasher {
     }
 
     /// Change the baud rate of the connection.
-    pub fn change_baud(&mut self, speed: u32) -> Result<(), Error> {
-        debug!("Change baud to: {}", speed);
+    pub fn change_baud(&mut self, baud: u32) -> Result<(), Error> {
+        debug!("Change baud to: {}", baud);
 
         let prior_baud = match self.use_stub {
-            true => self.connection.baud()?,
+            true => self.connection.baud(),
             false => 0,
         };
 
@@ -1152,7 +1143,7 @@ impl Flasher {
         //
         // The ROM code thinks it uses a 40 MHz XTAL. Recompute the baud rate in order
         // to trick the ROM code to set the correct baud rate for a 26 MHz XTAL.
-        let mut new_baud = speed;
+        let mut new_baud = baud;
         if self.chip == Chip::Esp32c2 && !self.use_stub && xtal_freq == XtalFrequency::_26Mhz {
             new_baud = new_baud * 40 / 26;
         }
@@ -1164,24 +1155,13 @@ impl Flasher {
                     prior_baud,
                 })
             })?;
-        self.connection.set_baud(speed)?;
+        self.connection.set_baud(baud)?;
         sleep(Duration::from_secs_f32(0.05));
         self.connection.flush()?;
 
         Ok(())
     }
 
-    /// Convert the [Flasher] into a [Port] instance.
-    pub fn into_serial(self) -> Port {
-        self.connection.into_serial()
-    }
-
-    /// Get the USB VID of the connected device.
-    pub fn usb_pid(&self) -> u16 {
-        self.connection.usb_pid()
-    }
-
-    /// Erase a region of flash specified by offset and size.
     pub fn erase_region(&mut self, offset: u32, size: u32) -> Result<(), Error> {
         debug!("Erasing region of 0x{:x}B at 0x{:08x}", size, offset);
 
@@ -1363,6 +1343,11 @@ impl Flasher {
 
         Ok(())
     }
+
+    /// Consume self and return the underlying connection.
+    pub fn into_connection(self) -> Connection {
+        self.connection
+    }
 }
 
 #[cfg(feature = "serialport")]
@@ -1381,24 +1366,6 @@ fn security_info(connection: &mut Connection, use_stub: bool) -> Result<Security
             ))
         }
     })
-}
-
-#[cfg(feature = "serialport")]
-fn detect_chip(connection: &mut Connection, use_stub: bool) -> Result<Chip, Error> {
-    match security_info(connection, use_stub) {
-        Ok(info) if info.chip_id.is_some() => {
-            let chip_id = info.chip_id.unwrap() as u16;
-            let chip = Chip::try_from(chip_id)?;
-
-            Ok(chip)
-        }
-        _ => {
-            let magic = connection.read_reg(CHIP_DETECT_MAGIC_REG_ADDR)?;
-            let chip = Chip::from_magic(magic)?;
-
-            Ok(chip)
-        }
-    }
 }
 
 #[cfg(feature = "serialport")]